refactor(feedback): extract reportable-field builder to a shared, pure module
The lens detail page built its spec-table groups and the report dialog's
field list inline. Extract that into resolveLensSpecAndFields(lens, locale,
t, tPricing) — pure in its inputs — so the same field list can be
reconstructed server-side from a lensId (needed for the upcoming document-flow
mobile feedback page) and never drifts from the rendered spec table.

Also moves the shared Feedback types to @/lib/feedback so lib code can import
them without a component → lib cycle. No behavior change: the detail spec
table and report dialog field picker render identically (verified).
